Changed some macros to explicitly specify the integer types.
This was causing a compile error. As I compiled using a new version
of the gcc cross-compiler, it was implicitly assuming signed integer
types where I assumed the unsigned 64-bit type. Now the raw integers
are explicitly typed to "size_t", the u64 type.

I should be careful in the future to explicitly specify this type.

+++ lib/memory.cpp
@@ -164,7 +164,7 @@ void free(void * const location) {
   size_t size = allocations[((__u64)location - (__u64)start) / BUDDY_PAGE_SIZE(0)];
 
   if (size <= BUDDY_MAX_PAGE_SIZE) {
-    for (int i = 0; i < BUDDY_MAX_ORDER; i++) {
+    for (size_t i = 0; i < BUDDY_MAX_ORDER; i++) {
       if (size <= BUDDY_PAGE_SIZE(i)) {
 	insertFreePage(location, i);
         compactFromOrder(i);
